+++ CHANGELOG.md
@@ -2,8 +2,25 @@
 
 All notable changes to the "api-test-builder" extension will be documented in this file.
 
-## [0.0.1] - 2014-05-31
+## [0.0.3] - 2024-12-03
 
-### Added
+**Added**
+
+- Swagger to Playwright (Java) - beta
+- Swagger to Playwright (Python) - beta
+- Swagger to Playwright (.Net) - beta
+
+## [0.0.2] - 2024-12-01
+
+**Added**
+
+- Swagger 2.0 version support
+
+## [0.0.1] - 2024-11-29
+
+**Added**
 
 - Initial release
+- OpenAPI 3.0.x version support
+- Swagger to Cypress (JavaScript)
+- Swagger to Playwright (JavaScript)